+++ foundation/src/main/scala/cocoa/foundation/NSMutableArray.scala
@@ -3,6 +3,8 @@ package cocoa.foundation
 
 import scala.language.experimental.macros
 import scalanative.native._
+import scalanative.unsafe._
+import scalanative.unsigned._
 import objc._
 
 @ObjC
@@ -59,7 +61,7 @@ object NSMutableArray extends NSMutableArrayClass {
   // TODO: use Iterable instead of Seq
   def arrayWithObjects[T<:NSObject](objects: Seq[T]): NSMutableArray[T] = Zone { implicit z =>
     val count = objects.size
-    val array = stackalloc[id]( sizeof[id] * count)
+    val array = stackalloc[id]( sizeof[id] * count.toUInt)
     for(i<-0 until count)
       !(array + i) = objects(i).toPtr
     arrayWithObjects_count_(array,count.toULong)
